requirements.txt
python-dateutil==2.9.0.post0
pandas==2.3.3
openpyxl==3.1.5
